Massive code cleanup and refactor
* Added missing implicit casts
* Added missing annotations
* Added missing PendingIntent flags
* Addressed possible NPEs
* Removed no-op build version checks
* Removed redundant code

+++ app/src/main/java/com/kabouzeid/gramophone/helper/StackBlur.java
@@ -153,9 +153,9 @@ public class StackBlur {
                     if (i <= wm) src_i += 1;
                     stack_i = i + radius;
                     stack[stack_i] = src[src_i];
-                    sum_r += ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
-                    sum_g += ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
-                    sum_b += (src[src_i] & 0xff) * (radius + 1 - i);
+                    sum_r += (long) ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
+                    sum_g += (long) ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
+                    sum_b += (long) (src[src_i] & 0xff) * (radius + 1 - i);
                     sum_in_r += ((src[src_i] >>> 16) & 0xff);
                     sum_in_g += ((src[src_i] >>> 8) & 0xff);
                     sum_in_b += (src[src_i] & 0xff);
@@ -242,9 +242,9 @@ public class StackBlur {
 
                     stack_i = i + radius;
                     stack[stack_i] = src[src_i];
-                    sum_r += ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
-                    sum_g += ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
-                    sum_b += (src[src_i] & 0xff) * (radius + 1 - i);
+                    sum_r += (long) ((src[src_i] >>> 16) & 0xff) * (radius + 1 - i);
+                    sum_g += (long) ((src[src_i] >>> 8) & 0xff) * (radius + 1 - i);
+                    sum_b += (long) (src[src_i] & 0xff) * (radius + 1 - i);
                     sum_in_r += ((src[src_i] >>> 16) & 0xff);
                     sum_in_g += ((src[src_i] >>> 8) & 0xff);
                     sum_in_b += (src[src_i] & 0xff);
